Enhance App Parameters Management with VAT and Discount Features
- Updated the UpdateAppParameters command and handler to include new properties for VatAmount and DiscountPercentage, improving financial data handling.
- Modified the UpdateAppParametersEndpoint to support the new parameters, ensuring they are processed correctly during updates.
- Enhanced the EditAppParameters component to allow user input for VAT and discount values, improving user experience and flexibility.
- Updated relevant database tables and configurations to accommodate the new decimal precision for financial calculations.
- Adjusted unit tests to validate the new parameters, ensuring robust testing coverage for the updated functionality.

+++ src/TunNetCom.SilkRoadErp.Sales.Contracts/AppParameters/GetAppParametersResponse.cs
@@ -42,8 +42,8 @@ public class GetAppParametersResponse
     public double PourcentageRetenu { get; set; }
 
     [JsonPropertyName("discountPercentage")]
-    public double DiscountPercentage { get; set; }
+    public decimal DiscountPercentage { get; set; }
 
     [JsonPropertyName("vatAmount")]
-    public double VatAmount { get; set; }
+    public decimal VatAmount { get; set; }
 }
